Create designated directory for downloading and organize functions into respective directories and files
